+++ opt/luarocks/luarocks/repos.lua
@@ -1,6 +1,8 @@
 
 --- Functions for managing the repository on disk.
-module("luarocks.rep", package.seeall)
+--module("luarocks.repos", package.seeall)
+local repos = {}
+package.loaded["luarocks.repos"] = repos
 
 local fs = require("luarocks.fs")
 local path = require("luarocks.path")
@@ -14,7 +16,7 @@ local deps = require("luarocks.deps")
 -- @param name string: a package name.
 -- @return table or nil: An array of strings listing installed
 -- versions of a package, or nil if none is available.
-function get_versions(name)
+local function get_installed_versions(name)
    assert(type(name) == "string")
    
    local dirs = fs.list_dir(path.versions_dir(name))
@@ -27,7 +29,7 @@ end
 -- @param version string: package version in string format
 -- @return boolean: true if a package is installed,
 -- false otherwise.
-function is_installed(name, version)
+function repos.is_installed(name, version)
    assert(type(name) == "string")
    assert(type(version) == "string")
       
@@ -53,22 +55,6 @@ local function recurse_rock_manifest_tree(file_tree, action)
    return do_recurse_rock_manifest_tree(file_tree, "", "")
 end
 
-local function store_package_data(result, name, sub, prefix)
-   assert(type(result) == "table")
-   assert(type(name) == "string")
-   assert(type(sub) == "table" or type(sub) == "string")
-   assert(type(prefix) == "string")
-
-   if type(sub) == "table" then
-      for sname, ssub in pairs(sub) do
-         store_package_data(result, sname, ssub, prefix..name.."/")
-      end
-   elseif type(sub) == "string" then
-      local pathname = prefix..name
-      result[path.path_to_module(pathname)] = pathname
-   end
-end
-
 local function store_package_data(result, name, file_tree)
    if not file_tree then return end
    return recurse_rock_manifest_tree(file_tree, 
@@ -88,7 +74,7 @@ end
 -- in "foo.bar" format and values are pathnames in architecture-dependent
 -- "foo/bar.so" format. If no modules are found or if package or version
 -- are invalid, an empty table is returned.
-function package_modules(package, version)
+function repos.package_modules(package, version)
    assert(type(package) == "string")
    assert(type(version) == "string")
 
@@ -107,7 +93,7 @@ end
 -- as strings and values are pathnames in architecture-dependent
 -- ".../bin/foo" format. If no modules are found or if package or version
 -- are invalid, an empty table is returned.
-function package_commands(package, version)
+function repos.package_commands(package, version)
    assert(type(package) == "string")
    assert(type(version) == "string")
 
@@ -123,7 +109,7 @@ end
 -- @param version string: version of an installed rock
 -- @return boolean: returns true if rock contains platform-specific
 -- binary executables, or false if it is a pure-Lua rock.
-function has_binaries(name, version)
+function repos.has_binaries(name, version)
    assert(type(name) == "string")
    assert(type(version) == "string")
 
@@ -139,7 +125,7 @@ function has_binaries(name, version)
    return false
 end
 
-function run_hook(rockspec, hook_name)
+function repos.run_hook(rockspec, hook_name)
    assert(type(rockspec) == "table")
    assert(type(hook_name) == "string")
 
@@ -147,6 +133,11 @@ function run_hook(rockspec, hook_name)
    if not hooks then
       return true
    end
+   
+   if cfg.hooks_enabled == false then
+      return nil, "This rockspec contains hooks, which are blocked by the 'hooks_enabled' setting in your LuaRocks configuration."
+   end
+   
    if not hooks.substituted_variables then
       util.variable_substitutions(hooks, rockspec.variables)
       hooks.substituted_variables = true
@@ -161,22 +152,16 @@ function run_hook(rockspec, hook_name)
    return true
 end
 
-local function install_binary(source, target)
+local function install_binary(source, target, name, version)
    assert(type(source) == "string")
    assert(type(target) == "string")
-
-   local match = source:match("%.lua$")
-   local file, ok, err
-   if not match then
-      file = io.open(source)
-   end
-   if match or (file and file:read():match("^#!.*lua.*")) then
-      ok, err = fs.wrap_script(source, target)
+   
+   if fs.is_lua(source) then
+      repos.ok, repos.err = fs.wrap_script(source, target, name, version)
    else
-      ok, err = fs.copy_binary(source, target)
+      repos.ok, repos.err = fs.copy_binary(source, target)
    end
-   if file then file:close() end
-   return ok, err
+   return repos.ok, repos.err
 end
 
 local function resolve_conflict(target, deploy_dir, name, version)
@@ -186,7 +171,8 @@ local function resolve_conflict(target, deploy_dir, name, version)
    end
    if name ~= cname or deps.compare_versions(version, cversion) then
       local versioned = path.versioned_name(target, deploy_dir, cname, cversion)
-      fs.make_dir(dir.dir_name(versioned))
+      local ok, err = fs.make_dir(dir.dir_name(versioned))
+      if not ok then return nil, err end
       fs.move(target, versioned)
       return target
    else
@@ -194,7 +180,7 @@ local function resolve_conflict(target, deploy_dir, name, version)
    end
 end
 
-function should_wrap_bin_scripts(rockspec)
+function repos.should_wrap_bin_scripts(rockspec)
    assert(type(rockspec) == "table")
 
    if cfg.wrap_bin_scripts ~= nil then
@@ -206,12 +192,13 @@ function should_wrap_bin_scripts(rockspec)
    return true
 end
 
-function deploy_files(name, version, wrap_bin_scripts)
+function repos.deploy_files(name, version, wrap_bin_scripts)
    assert(type(name) == "string")
    assert(type(version) == "string")
    assert(type(wrap_bin_scripts) == "boolean")
 
-   local function deploy_file_tree(file_tree, source_dir, deploy_dir, move_fn)
+   local function deploy_file_tree(file_tree, path_fn, deploy_dir, move_fn)
+      local source_dir = path_fn(name, version)
       if not move_fn then
          move_fn = fs.move
       end
@@ -223,15 +210,21 @@ function deploy_files(name, version, wrap_bin_scripts)
             if fs.exists(target) then
                local new_target, err = resolve_conflict(target, deploy_dir, name, version)
                if err == "untracked" then
-                  fs.delete(target)
+                  local backup = target
+                  repeat
+                     backup = backup.."~"
+                  until not fs.exists(backup) -- slight race condition here, but shouldn't be a problem.
+                  util.printerr("Warning: "..target.." is not tracked by this installation of LuaRocks. Moving it to "..backup)
+                  fs.move(target, backup)
                elseif err then
                   return nil, err.." Cannot install new version."
                else
                   target = new_target
                end
-	    end
-            fs.make_dir(dir.dir_name(target))
-            ok, err = move_fn(source, target)
+            end
+            ok, err = fs.make_dir(dir.dir_name(target))
+            if not ok then return nil, err end
+            ok, err = move_fn(source, target, name, version)
             fs.remove_dir_tree_if_empty(dir.dir_name(source))
             if not ok then return nil, err end
             return true
@@ -244,45 +237,66 @@ function deploy_files(name, version, wrap_bin_scripts)
    local ok, err = true
    if rock_manifest.bin then
       local move_bin_fn = wrap_bin_scripts and install_binary or fs.copy_binary
-      ok, err = deploy_file_tree(rock_manifest.bin, path.bin_dir(name, version), cfg.deploy_bin_dir, move_bin_fn)
+      ok, err = deploy_file_tree(rock_manifest.bin, path.bin_dir, cfg.deploy_bin_dir, move_bin_fn)
    end
    if ok and rock_manifest.lua then
-      ok, err = deploy_file_tree(rock_manifest.lua, path.lua_dir(name, version), cfg.deploy_lua_dir)
+      ok, err = deploy_file_tree(rock_manifest.lua, path.lua_dir, cfg.deploy_lua_dir)
    end
    if ok and rock_manifest.lib then
-      ok, err = deploy_file_tree(rock_manifest.lib, path.lib_dir(name, version), cfg.deploy_lib_dir)
+      ok, err = deploy_file_tree(rock_manifest.lib, path.lib_dir, cfg.deploy_lib_dir)
    end
    return ok, err
+end
+
+local function delete_suffixed(filename, suffix)
+   local filenames = { filename }
+   if suffix and suffix ~= "" then filenames = { filename..suffix, filename } end
+   for _, name in ipairs(filenames) do
+      if fs.exists(name) then
+         fs.delete(name)
+         if fs.exists(name) then
+            return nil, "Failed deleting "..name, "fail"
+         end
+         return true, name
+      end
+   end
+   return false, "File not found", "not found"
 end
 
 --- Delete a package from the local repository.
 -- Version numbers are compared as exact string comparison.
 -- @param name string: name of package
 -- @param version string: package version in string format
-function delete_version(name, version)
+-- @param quick boolean: do not try to fix the versioned name
+-- of another version that provides the same module that
+-- was deleted. This is used during 'purge', as every module
+-- will be eventually deleted.
+function repos.delete_version(name, version, quick)
    assert(type(name) == "string")
    assert(type(version) == "string")
 
-   local function delete_deployed_file_tree(file_tree, deploy_dir)
+   local function delete_deployed_file_tree(file_tree, deploy_dir, suffix)
       return recurse_rock_manifest_tree(file_tree, 
          function(parent_path, parent_module, file)
             local target = dir.path(deploy_dir, parent_path, file)
             local versioned = path.versioned_name(target, deploy_dir, name, version)
-            if fs.exists(versioned) then
-               local ok = fs.delete(versioned)
+            local ok, name, err = delete_suffixed(versioned, suffix)
+            if ok then
                fs.remove_dir_tree_if_empty(dir.dir_name(versioned))
-               if not ok then return nil, "Failed deleting "..versioned end
-            else
-               local ok = fs.delete(target)
+               return true
+            end
+            if err == "fail" then return nil, name end
+            ok, name, err = delete_suffixed(target, suffix)
+            if err == "fail" then return nil, name end
+            if not quick then
                local next_name, next_version = manif.find_next_provider(target)
                if next_name then
-                  local versioned = path.versioned_name(target, deploy_dir, next_name, next_version)
-                  fs.move(versioned, target)
+                  local versioned = path.versioned_name(name, deploy_dir, next_name, next_version)
+                  fs.move(versioned, name)
                   fs.remove_dir_tree_if_empty(dir.dir_name(versioned))
                end
-               fs.remove_dir_tree_if_empty(dir.dir_name(target))
-               if not ok then return nil, "Failed deleting "..target end
             end
+            fs.remove_dir_tree_if_empty(dir.dir_name(target))
             return true
          end
       )
@@ -295,7 +309,7 @@ function delete_version(name, version)
    
    local ok, err = true
    if rock_manifest.bin then
-      ok, err = delete_deployed_file_tree(rock_manifest.bin, cfg.deploy_bin_dir)
+      ok, err = delete_deployed_file_tree(rock_manifest.bin, cfg.deploy_bin_dir, cfg.wrapper_suffix)
    end
    if ok and rock_manifest.lua then
       ok, err = delete_deployed_file_tree(rock_manifest.lua, cfg.deploy_lua_dir)
@@ -306,8 +320,10 @@ function delete_version(name, version)
    if err then return nil, err end
 
    fs.delete(path.install_dir(name, version))
-   if not get_versions(name) then
+   if not get_installed_versions(name) then
       fs.delete(dir.path(cfg.rocks_dir, name))
    end
    return true
 end
+
+return repos
